commit 2017/05/11 12.22 PM - new changes apply for user login

+++ src/main/java/com/leaf/school/dto/common/DataTableResponseDTO.java
@@ -6,10 +6,10 @@ import java.util.List;
 /**
  * @Author : Rusiru De Silva
  */
-public class DataTableResponseDTO implements Serializable{
+public class DataTableResponseDTO implements Serializable {
     private Integer draw;
-    private Integer recordsTotal;
-    private Integer recordsFiltered;
+    private Long recordsTotal;
+    private Long recordsFiltered;
     private List<?> data;
 
     public Integer getDraw() {
@@ -20,19 +20,19 @@ public class DataTableResponseDTO implements Serializable{
         this.draw = draw;
     }
 
-    public Integer getRecordsTotal() {
+    public Long getRecordsTotal() {
         return recordsTotal;
     }
 
-    public void setRecordsTotal(Integer recordsTotal) {
+    public void setRecordsTotal(Long recordsTotal) {
         this.recordsTotal = recordsTotal;
     }
 
-    public Integer getRecordsFiltered() {
+    public Long getRecordsFiltered() {
         return recordsFiltered;
     }
 
-    public void setRecordsFiltered(Integer recordsFiltered) {
+    public void setRecordsFiltered(Long recordsFiltered) {
         this.recordsFiltered = recordsFiltered;
     }
