+++ src/main/java/com/leaf/school/entity/StatusCategoryEntity.java
@@ -2,25 +2,27 @@ package com.leaf.school.entity;
 /**
  * @Author : Rusiru De Silva
  */
+
 import javax.persistence.*;
 
 
 @Entity
 @Table(name = "master_data.status_category")
 public class StatusCategoryEntity {
-    private Integer id;
+    private Long id;
     private String code;
     private String description;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
+
     @Column(name = "code", nullable = false)
     public String getCode() {
         return code;
@@ -29,6 +31,7 @@ public class StatusCategoryEntity {
     public void setCode(String code) {
         this.code = code;
     }
+
     @Column(name = "description", nullable = false)
     public String getDescription() {
         return description;
